Moves can stuff out of the bridge.
