+++ CAN-D/Src/can.c
@@ -8,27 +8,35 @@
 
 /* Includes ------------------------------------------------------------------*/
 #include "can.h"
-#include "bridge.h"
 #include "fatfs.h"
+#include "stm32302c_custom.h"
 #include "usbd_cdc_if.h"
-
-osPoolDef(CANTxPool, 8, CANTxMessage);
 
 /* Private typedef -----------------------------------------------------------*/
 
 /* Private define ------------------------------------------------------------*/
 
 /* Private variables ---------------------------------------------------------*/
+osThreadId bridgeConfigTaskHandle;
+osThreadId CANMonitorTaskHandle;
+osThreadId CANTransmitTaskHandle;
+osMessageQId CANRxQueueHandle;
+osMessageQId CANTxQueueHandle;
+APP_ConfigType mAppConfiguration = { 0 };
+osPoolDef(CANTxPool, 8, CANTxMessage);
+osPoolId CANTxPool;
 
 /* Private macro -------------------------------------------------------------*/
 
 /* Exported variables --------------------------------------------------------*/
 CAN_HandleTypeDef hcan;
-extern APP_ConfigType mAppConfiguration;
 extern USBD_HandleTypeDef hUsbDeviceFS;
 extern osPoolId CANTxPool;
 
 /* Private function prototypes -----------------------------------------------*/
+void APP_CAN_LogButtonTask(void const* argument);
+void APP_CAN_MonitorTask(void const* argument);
+void APP_CAN_TransmitTask(void const* argument);
 
 /* Exported functions --------------------------------------------------------*/
 /* CAN init function */
@@ -54,7 +62,30 @@ void MX_CAN_Init(void)
     if (CANTxPool == NULL) {
         Error_Handler();
     }
+
     // TODO: configure CAN reception filters using HAL_CAN_ConfigFilter()
+    mAppConfiguration.SDStorage = APP_DISABLE;
+    mAppConfiguration.USBStream = APP_DISABLE;
+    mAppConfiguration.USBTransfer = APP_DISABLE;
+    mAppConfiguration.CANTransmit = APP_DISABLE;
+}
+
+void APP_CAN_InitTasks(void)
+{
+    osThreadDef(bridgeConfigTask, APP_CAN_LogButtonTask, osPriorityNormal, 0, 128);
+    bridgeConfigTaskHandle = osThreadCreate(osThread(bridgeConfigTask), NULL);
+
+    osThreadDef(CANMonitorTask, APP_CAN_MonitorTask, osPriorityNormal, 0, 128);
+    CANMonitorTaskHandle = osThreadCreate(osThread(CANMonitorTask), NULL);
+
+    osThreadDef(CANTransmitTask, APP_CAN_TransmitTask, osPriorityNormal, 0, 128);
+    CANTransmitTaskHandle = osThreadCreate(osThread(CANTransmitTask), NULL);
+
+    osMessageQDef(CANRxQueue, 8, uint8_t);
+    CANRxQueueHandle = osMessageCreate(osMessageQ(CANRxQueue), NULL);
+
+    osMessageQDef(CANTxQueue, 16, CANTxMessage);
+    CANTxQueueHandle = osMessageCreate(osMessageQ(CANTxQueue), NULL);
 }
 
 void HAL_CAN_MspInit(CAN_HandleTypeDef* canHandle)
@@ -137,7 +168,7 @@ void HAL_CAN_RxFifo0MsgPendingCallback(CAN_HandleTypeDef* canHandle)
     }
 
     // Pass CAN RX data to the monitor thread
-    // @see APP_BRIDGE_CANMonitorTask()
+    // @see APP_CAN_MonitorTask()
     osMessagePut(CANRxQueueHandle, (uint32_t)rxData[0], 0);
 }
 
@@ -181,4 +212,98 @@ void APP_CAN_TransmitData(uint8_t* txData, CAN_TxHeaderTypeDef* header)
     osMessagePut(CANTxQueueHandle, (uint32_t)msg, 0);
 }
 
+/**
+  * @brief  Sets the APP Configuration
+  * @param  newConfig: the configuration to set
+  * @retval None
+  */
+void APP_CAN_SetConfiguration(APP_ConfigType newConfig)
+{
+    mAppConfiguration = newConfig;
+}
+
 /* Private functions ---------------------------------------------------------*/
+/**
+  * @brief  Function implementing the APP_CAN_LogButtonTask thread.
+  * @param  argument: Not used
+  * @retval None
+  */
+void APP_CAN_LogButtonTask(void const* argument)
+{
+    for (;;) {
+        // Start/Stop the CAN module on button press
+        if (BSP_PB_GetState(BUTTON_LOG) == GPIO_PIN_SET) {
+            APP_CAN_StartStop();
+        }
+
+        osDelay(1);
+    }
+}
+
+/**
+  * @brief  Function implementing the APP_CAN_MonitorTask thread.
+  *         Monitors incoming CAN data.
+  * @retval None
+  */
+void APP_CAN_MonitorTask(void const* argument)
+{
+    osEvent event;
+
+    for (;;) {
+        // Pend on any CAN Rx data
+        event = osMessageGet(CANRxQueueHandle, 0);
+        if (event.status == osEventMessage) {
+            if (mAppConfiguration.SDStorage == APP_ENABLE) {
+                // Write data to SD card
+                APP_FATFS_WriteSD((const uint8_t*)event.value.p, 8, "CAN_data.log");
+            }
+            if (mAppConfiguration.USBStream == APP_ENABLE) {
+                // Fill the USB TX buffer with the CAN data
+                while (CDC_Transmit_FS((uint8_t*)event.value.p, (uint16_t)8) == 1) {
+                    // USB TX State is BUSY. Wait for it to be free.
+                }
+            }
+        }
+        osDelay(1);
+    }
+}
+
+/**
+  * @brief  Function implementing the APP_CAN_TransmitTask thread.
+  *         Send outgoing CAN data.
+  * @retval None
+  */
+void APP_CAN_TransmitTask(void const* argument)
+{
+    osEvent event;
+    CANTxMessage* msg;
+    uint32_t mailbox;
+
+    for (;;) {
+        // Pend on any CAN Tx data
+        event = osMessageGet(CANTxQueueHandle, 0);
+        if (event.status == osEventMessage) {
+            msg = event.value.p;
+            if (mAppConfiguration.CANTransmit == APP_ENABLE) {
+                if (HAL_CAN_GetTxMailboxesFreeLevel(msg->handle) > 0) {
+                    /* Default to sending on MAILBOX0 */
+                    mailbox = CAN_TX_MAILBOX0;
+                    /* Check Tx Mailbox 1 status */
+                    if ((msg->handle->Instance->TSR & CAN_TSR_TME1) != 0U) {
+                        mailbox = CAN_TX_MAILBOX1;
+                    }
+                    /* Check Tx Mailbox 2 status */
+                    else if ((msg->handle->Instance->TSR & CAN_TSR_TME2) != 0U) {
+                        mailbox = CAN_TX_MAILBOX2;
+                    }
+                    HAL_CAN_AddTxMessage(msg->handle, msg->header, msg->data, (uint32_t*)mailbox);
+                } else {
+                    // Dropped a message!
+                }
+            }
+            osPoolFree(CANTxPool, msg);
+        }
+
+        osDelay(1);
+    }
+}
